+++ frontend/src/pages/protocol-token/[stablecoin].vue
@@ -31,11 +31,9 @@
 								{{ pltTokenDataRef?.name }}
 							</p>
 						</div>
-						<div class="flex justify-between pt-4 gap-4">
-							<p class="text-xl text-theme-faded flex-shrink-0">Symbol</p>
-							<p
-								class="font-bold text-xl text-theme-interactive text-right break-words overflow-wrap-anywhere max-w-[60%]"
-							>
+						<div class="flex justify-between pt-4">
+							<p class="text-xl text-theme-faded">Token ID</p>
+							<p class="font-bold text-xl text-theme-interactive">
 								{{ pltTokenDataRef?.tokenId }}
 							</p>
 						</div>